+++ src/pages/ChallengeList/template/index.tsx
@@ -1,5 +1,5 @@
 import { ChallengeListData } from 'apis';
-import { ChallengeDetailCard } from 'components/molecules';
+import { ChallengeDetailCard } from 'components/organisms';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useRecoilValue } from 'recoil';
 import { userStatusState } from 'stores/user';
