+++ src/pages/ChallengeList/template/style.ts
@@ -1,0 +1,11 @@
+import Styled from 'styled-components';
+import { theme } from 'styled-tools';
+
+export const ChallengeCardWrapper = Styled.div`
+p{
+  margin : 100px 0 100px 0;
+  width : 100%;
+  text-align  :center;
+  ${theme('font.body3')};
+}
+`;
